% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{PostToSlack}
\alias{PostToSlack}
\title{Post an arbitrary message to Slack}
\usage{
PostToSlack(channelName, messageToPost, botName = NULL,
  webhookURL = Sys.getenv("SLACK_WEBHOOK"))
}
\arguments{
\item{channelName}{A string of the form \code{"@username"} or \code{"#channelname"}. When using
\code{"@username"}, log messages will be sent through Slackbot to that user. When using
\code{"#channelname"}, log messages will be posted to that channel as a bot.}

\item{messageToPost}{A string of the message to post.}

\item{botName}{A string of the name of the bot that posts. Default is NULL, which means to use
the default bot name that was set up with the Incoming Webhook. See References.}

\item{webhookURL}{A string of the Slack webhook URL of the form
\code{"https://hooks.slack.com/services/XXX/YYY/ZZZ"}. Default reads the environment variable
\code{SLACK_WEBHOOK}. See References.}
}
\value{
Result of the \code{\link[httr]{POST}} call.
}
\description{
Posts a message to Slack using Incoming Webhooks
}
\references{
\href{https://get.slack.help/hc/en-us/articles/115005265063-Incoming-WebHooks-for-Slack}{Slack
documentation on setting up an Incoming Webhook}
}
